+++ src/components/People/Alex/Alex.jsx
@@ -1,0 +1,6 @@
+export const alex = {
+  name: 'Alex',
+  age: 25,
+  sex: 'm',
+  isMarried: false,
+};
