+++ src/components/People/Alex/index.jsx
@@ -1,0 +1,1 @@
+export * from './Alex';
