+++ src/components/People/Misha/Misha.jsx
@@ -1,0 +1,7 @@
+export const misha = {
+  name: 'Misha',
+  age: 37,
+  sex: 'm',
+  isMarried: true,
+  partnerName: 'Natasha',
+};
