+++ src/components/People/Misha/index.jsx
@@ -1,0 +1,1 @@
+export * from './Misha';
