+++ src/components/People/Olya/Olya.jsx
@@ -1,0 +1,6 @@
+export const olya = {
+  name: 'Olya',
+  sex: 'f',
+  isMarried: true,
+  partnerName: 'Maksym',
+};
