+++ src/components/People/Olya/index.jsx
@@ -1,0 +1,1 @@
+export * from './Olya';
